Fix tools - Change haders in the tools - Create fheroe2/agg/decode.cpp which has logic for ICN decoding. It is separate file, to not take all sound related functions in the tools which make it complicated to build. - Add VC projects for building of each tool
