+++ src/tools/icn2img.cpp
@@ -26,66 +26,20 @@
 #include <iostream>
 #include <sstream>
 
-#include "engine.h"
+#include "serialize.h"
 #include "palette_h2.h"
 #include "system.h"
+#include "image.h"
+#include "image_tool.h"
+#include <SDL.h>
+#include "decode.h"
+#include "engine.h"
 
-struct icnheader
-{
-    s16 offsetX;
-    s16 offsetY;
-    u16 width;
-    u16 height;
-    u8 type; // type of sprite : 0 = Normal, 32 = Monochromatic shape
-    u32 offsetData;
+#if defined( _MSC_VER )
+#undef main
+#endif
 
-    icnheader()
-        : offsetX( 0 )
-        , offsetY( 0 )
-        , width( 0 )
-        , height( 0 )
-        , type( 0 )
-        , offsetData( 0 )
-    {}
-};
-
-StreamBase & operator>>( StreamBase & sb, icnheader & st )
-{
-    return sb >> st.offsetX >> st.offsetY >> st.width >> st.height >> st.type >> st.offsetData;
-}
-
-void SpriteDrawICNv1( Surface & sf, const std::vector<u8> &, bool debug );
-void SpriteDrawICNv2( Surface & sf, const std::vector<u8> &, bool debug );
-
-namespace H2Palette
-{
-    std::vector<SDL_Color> pal_colors;
-
-    void Init( void )
-    {
-        // load palette
-        u32 ncolors = ARRAY_COUNT( kb_pal ) / 3;
-        pal_colors.reserve( ncolors );
-
-        for ( u32 ii = 0; ii < ncolors; ++ii ) {
-            u32 index = ii * 3;
-            SDL_Color cols;
-
-            cols.r = kb_pal[index] << 2;
-            cols.g = kb_pal[index + 1] << 2;
-            cols.b = kb_pal[index + 2] << 2;
-
-            pal_colors.push_back( cols );
-        }
-
-        Surface::SetDefaultPalette( &pal_colors[0], pal_colors.size() );
-    }
-
-    RGBA GetColor( u32 index )
-    {
-        return index < pal_colors.size() ? RGBA( pal_colors[index].r, pal_colors[index].g, pal_colors[index].b ) : RGBA( 0, 0, 0 );
-    }
-}
+StreamBase & operator>>( StreamBase & st, fheroes2::AGG::ICNHeader & icn );
 
 int main( int argc, char ** argv )
 {
@@ -143,18 +97,17 @@ int main( int argc, char ** argv )
     }
 
     SDL::Init();
-    H2Palette::Init();
 
     fs << "<?xml version=\"1.0\" ?>" << std::endl << "<icn name=\"" << shortname << ".icn\" count=\"" << count_sprite << "\">" << std::endl;
 
     u32 save_pos = sf.tell();
 
-    std::vector<icnheader> headers( count_sprite );
+    std::vector<fheroes2::AGG::ICNHeader> headers( count_sprite );
     for ( int ii = 0; ii < count_sprite; ++ii )
         sf >> headers[ii];
 
     for ( int ii = 0; ii < count_sprite; ++ii ) {
-        const icnheader & head = headers[ii];
+        const fheroes2::AGG::ICNHeader & head = headers[ii];
 
         u32 data_size = ( ii + 1 != count_sprite ? headers[ii + 1].offsetData - head.offsetData : total_size - head.offsetData );
         sf.seek( save_pos + head.offsetData );
@@ -162,17 +115,8 @@ int main( int argc, char ** argv )
         std::vector<u8> buf = sf.getRaw( data_size );
 
         if ( buf.size() ) {
-            Surface surf( Size( head.width, head.height ), /*false*/ true ); // accepting transparency
-
-            const RGBA clkey = RGBA( 0xFF, 0, 0xFF );
-            surf.Fill( clkey );
-            surf.SetColorKey( clkey );
-            surf.Fill( RGBA( 0, 0, 0, 0xFF ) ); // filling with transparent color
-
-            if ( 0x20 == head.type )
-                SpriteDrawICNv2( surf, buf, debug );
-            else
-                SpriteDrawICNv1( surf, buf, debug );
+            fheroes2::Sprite image; // accepting transparency
+            fheroes2::AGG::DecodeICNSprite( image, head, &buf[0], data_size );
 
             std::ostringstream os;
             os << std::setw( 3 ) << std::setfill( '0' ) << ii;
@@ -187,9 +131,9 @@ int main( int argc, char ** argv )
             dstfile += ".png";
             shortdstfile += ".png";
 #endif
-            surf.Save( dstfile.c_str() );
-            fs << " <sprite index=\"" << ii + 1 << "\" name=\"" << shortdstfile.c_str() << "\" ox=\"" << head.offsetX << "\" oy=\"" << head.offsetY << "\"/>"
-               << std::endl;
+            if ( !fheroes2::Save( image, dstfile, 0 ) )
+                fs << " <sprite index=\"" << ii + 1 << "\" name=\"" << shortdstfile.c_str() << "\" ox=\"" << head.offsetX << "\" oy=\"" << head.offsetY << "\"/>"
+                   << std::endl;
         }
     }
 
@@ -200,157 +144,4 @@ int main( int argc, char ** argv )
 
     SDL::Quit();
     return EXIT_SUCCESS;
-}
-
-void SpriteDrawICNv1( Surface & sf, const std::vector<u8> & buf, bool debug )
-{
-    const u8 * cur = &buf[0];
-    const u32 size = buf.size();
-    const u8 * max = cur + size;
-
-    u8 c = 0;
-    u16 x = 0;
-    u16 y = 0;
-
-    RGBA shadow = RGBA( 0, 0, 0, 0x40 );
-
-    // lock surface
-    while ( 1 ) {
-        if ( debug )
-            std::cerr << "CMD:"
-                      << "0x" << std::setw( 2 ) << std::setfill( '0' ) << std::hex << static_cast<int>( *cur );
-
-        // 0x00 - end line
-        if ( 0 == *cur ) {
-            ++y;
-            x = 0;
-            ++cur;
-        }
-        else
-            // 0x7F - count data
-            if ( 0x80 > *cur ) {
-            c = *cur;
-            ++cur;
-            while ( c-- && cur < max ) {
-                sf.DrawPoint( Point( x, y ), H2Palette::GetColor( *cur ) );
-                ++x;
-                ++cur;
-            }
-        }
-        else
-            // 0x80 - end data
-            if ( 0x80 == *cur ) {
-            if ( debug )
-                std::cerr << std::endl;
-
-            break;
-        }
-        else
-            // 0xBF - skip data
-            if ( 0xC0 > *cur ) {
-            x += *cur - 0x80;
-            ++cur;
-        }
-        else
-            // 0xC0 - shadow
-            if ( 0xC0 == *cur ) {
-            ++cur;
-            c = ( *cur % 4 ) ? *cur % 4 : *( ++cur );
-
-            while ( c-- ) {
-                sf.DrawPoint( Point( x, y ), shadow );
-                ++x;
-            }
-
-            ++cur;
-        }
-        else
-            // 0xC1
-            if ( 0xC1 == *cur ) {
-            ++cur;
-            c = *cur;
-            ++cur;
-            while ( c-- ) {
-                sf.DrawPoint( Point( x, y ), H2Palette::GetColor( *cur ) );
-                ++x;
-            }
-            ++cur;
-        }
-        else {
-            c = *cur - 0xC0;
-            ++cur;
-            while ( c-- ) {
-                sf.DrawPoint( Point( x, y ), H2Palette::GetColor( *cur ) );
-                ++x;
-            }
-            ++cur;
-        }
-
-        if ( cur >= max ) {
-            std::cerr << "out of range" << std::endl;
-            break;
-        }
-
-        if ( debug )
-            std::cerr << std::endl;
-    }
-}
-
-void SpriteDrawICNv2( Surface & sf, const std::vector<u8> & buf, bool debug )
-{
-    const u8 * cur = &buf[0];
-    const u32 size = buf.size();
-    const u8 * max = cur + size;
-
-    u8 c = 0;
-    u16 x = 0;
-    u16 y = 0;
-
-    RGBA shadow = RGBA( 0, 0, 0, 0xff );
-
-    // lock surface
-    while ( 1 ) {
-        if ( debug )
-            std::cerr << "CMD:"
-                      << "0x" << std::setw( 2 ) << std::setfill( '0' ) << std::hex << static_cast<int>( *cur );
-
-        // 0x00 - end line
-        if ( 0 == *cur ) {
-            ++y;
-            x = 0;
-            ++cur;
-        }
-        else
-            // 0x7F - count data
-            if ( 0x80 > *cur ) {
-            c = *cur;
-            while ( c-- ) {
-                sf.DrawPoint( Point( x, y ), shadow );
-                ++x;
-            }
-            ++cur;
-        }
-        else
-            // 0x80 - end data
-            if ( 0x80 == *cur ) {
-            if ( debug )
-                std::cerr << std::endl;
-
-            break;
-        }
-        else
-        // other - skip data
-        {
-            x += *cur - 0x80;
-            ++cur;
-        }
-
-        if ( cur >= max ) {
-            std::cerr << "out of range" << std::endl;
-            break;
-        }
-
-        if ( debug )
-            std::cerr << std::endl;
-    }
 }
