+++ src/tools/til2img.cpp
@@ -28,38 +28,15 @@
 #include <SDL.h>
 
 #include "engine.h"
+#include "serialize.h"
 #include "palette_h2.h"
 #include "system.h"
+#include "image.h"
+#include "image_tool.h"
 
-namespace H2Palette
-{
-    std::vector<SDL_Color> pal_colors;
-
-    void Init( void )
-    {
-        // load palette
-        u32 ncolors = ARRAY_COUNT( kb_pal ) / 3;
-        pal_colors.reserve( ncolors );
-
-        for ( u32 ii = 0; ii < ncolors; ++ii ) {
-            u32 index = ii * 3;
-            SDL_Color cols;
-
-            cols.r = kb_pal[index] << 2;
-            cols.g = kb_pal[index + 1] << 2;
-            cols.b = kb_pal[index + 2] << 2;
-
-            pal_colors.push_back( cols );
-        }
-
-        Surface::SetDefaultPalette( &pal_colors[0], pal_colors.size() );
-    }
-
-    RGBA GetColor( u32 index )
-    {
-        return index < pal_colors.size() ? RGBA( pal_colors[index].r, pal_colors[index].g, pal_colors[index].b ) : RGBA( 0, 0, 0 );
-    }
-}
+#if defined( _MSC_VER )
+#undef main
+#endif
 
 int main( int argc, char ** argv )
 {
@@ -78,7 +55,9 @@ int main( int argc, char ** argv )
     std::string prefix( argv[2] );
     std::string shortname( argv[1] );
 
+    bool debugMode = false;
     if ( shortname == "-d" ) {
+        debugMode = true;
     }
 
     shortname.replace( shortname.find( "." ), 4, "" );
@@ -93,15 +72,19 @@ int main( int argc, char ** argv )
     int count = sf.getLE16();
     int width = sf.getLE16();
     int height = sf.getLE16();
-    std::vector<u8> buf = sf.getRaw( size );
-
+    std::vector<u8> buf = sf.getRaw( width * height * count );
+    if ( debugMode ) {
+        std::cout << "Size of stream " << size << "(" << buf.size() << ")" << std::endl;
+        std::cout << "Count of images " << count << "(" << width << "," << height << ")" << std::endl; 
+    }
+    
     SDL::Init();
-    H2Palette::Init();
 
     for ( int cur = 0; cur < count; ++cur ) {
         u32 offset = width * height * cur;
         if ( offset < buf.size() ) {
-            Surface sf( &buf[offset], width, height, 1, false );
+            fheroes2::Image image( width, height );
+            memcpy( image.image(), &buf[offset], (size_t)width * height );
 
             std::ostringstream stream;
             stream << std::setw( 3 ) << std::setfill( '0' ) << cur;
@@ -112,7 +95,11 @@ int main( int argc, char ** argv )
 #else
             dstfile += ".png";
 #endif
-            if ( !sf.Save( dstfile.c_str() ) )
+            if ( debugMode ) {
+                std::cout << "Saving " << dstfile << std::endl;
+            }
+            
+            if ( !fheroes2::Save( image, dstfile, 0 ) )
                 std::cout << "error" << std::endl;
         }
     }
